requirements.txt
streamlit==1.34.0
python-docx==0.8.11
pypdf==4.2.0
nltk==3.8.1
wordcloud==1.9.3
matplotlib==3.8.3
spacy==3.7.4
requests==2.31.0
spacy-lookups-data==1.0.3
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
pyperclip==1.8.2
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
tqdm==4.66.2
reportlab==4.1.0
openai>=1.12.0
